requirements.txt
google-generativeai
requests
python-telegram-bot
APScheduler
PyYAML
